requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.1
django-jwt-extended==1.0.0
django-xmlrpc==0.1.8
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pypi-xmlrpc==2020.12.3
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
